+++ internal/cache/mem.go
@@ -1,28 +1,39 @@
 package cache
 
 import (
+	"context"
 	"time"
+
+	"github.com/go-redis/cache/v8"
 )
 
 type Memory struct {
-	M map[string]interface{}
+	C   *cache.Cache
+	TTL time.Duration
 }
 
-func NewMemory(expire, purge time.Duration) *Memory {
-	return &Memory{make(map[string]interface{})}
+func NewMemory(ttl time.Duration) *Memory {
+	return &Memory{cache.New(&cache.Options{
+		LocalCache: cache.NewTinyLFU(1000, ttl),
+	}), ttl}
 }
 
 func (m *Memory) Set(key string, val interface{}) error {
-	m.M[key] = val
-	return nil
+	ctx := context.Background()
+	return m.C.Set(&cache.Item{
+		Ctx:   ctx,
+		Key:   key,
+		Value: val,
+		TTL:   m.TTL,
+	})
 }
 
-func (m *Memory) Get(key string) (interface{}, bool) {
-	val, ok := m.M[key]
-	return val, ok
+func (m *Memory) Get(key string, value interface{}) error {
+	ctx := context.Background()
+	return m.C.Get(ctx, key, value)
 }
 
 func (m *Memory) Drop(key string) error {
-	delete(m.M, key)
-	return nil
+	ctx := context.Background()
+	return m.C.Delete(ctx, key)
 }