+++ internal/cache/named.go
@@ -15,8 +15,8 @@ func (c *NamedCache) Set(key string, val interface{}) error {
 	return c.c.Set(c.prefix+":"+key, val)
 }
 
-func (c *NamedCache) Get(key string) (interface{}, bool) {
-	return c.c.Get(c.prefix + ":" + key)
+func (c *NamedCache) Get(key string, val interface{}) error {
+	return c.c.Get(c.prefix+":"+key, val)
 }
 
 func (c *NamedCache) Drop(key string) error {