+++ internal/meta/cacheservice.go
@@ -1,29 +1,27 @@
 package meta
 
-import "go.uber.org/zap"
-
-type Cache interface {
-	Set(key string, val interface{}) error
-	Get(key string) (interface{}, bool)
-}
+import (
+	"github.com/2785/warframe-assistant/internal/cache"
+	"go.uber.org/zap"
+)
 
 type CacheService struct {
-	c Cache
+	c cache.Cache
 	l *zap.Logger
 
 	Service
 }
 
-func NewWithCache(s Service, c Cache, l *zap.Logger) *CacheService {
+func NewWithCache(s Service, c cache.Cache, l *zap.Logger) *CacheService {
 	return &CacheService{c, l, s}
 }
 
 func (s *CacheService) GetRoleRequirementForGuild(action string, gid string) (string, error) {
-	if val, ok := s.c.Get(gid + ":" + action); ok {
-		s, ok := val.(string)
-		if ok {
-			return s, nil
-		}
+	rid := ""
+
+	err := s.c.Get(gid+":"+action, &rid)
+	if err == nil {
+		return rid, nil
 	}
 
 	role_id, err := s.Service.GetRoleRequirementForGuild(action, gid)