+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 ## [Unreleased]
 
 
+<a name="v0.2.0"></a>
+## [v0.2.0] - 2021-06-13
+### Bug Fixes
+- empty role is now properly handled
+
+### Docs
+- add testing instructions to readme
+- add changelog
+
+### Features
+- add redis caching
+
+
 <a name="v0.1.0"></a>
 ## v0.1.0 - 2021-06-11
 ### Bug Fixes
@@ -18,4 +31,5 @@
 - add support for score events
 
 
-[Unreleased]: https://github.com/2785/warframe-assistant/compare/v0.1.0...HEAD
+[Unreleased]: https://github.com/2785/warframe-assistant/compare/v0.2.0...HEAD
+[v0.2.0]: https://github.com/2785/warframe-assistant/compare/v0.1.0...v0.2.0